moved csv_reader to a new common.py file to decrease dup code
