+++ iridauploader/parsers/common.py
@@ -1,0 +1,35 @@
+from os import path
+from csv import reader
+
+from iridauploader.parsers import exceptions
+
+
+def get_csv_reader(sample_sheet_file):
+
+    """
+    tries to create a csv.reader object which will be used to
+        parse through the lines in SampleSheet.csv
+    raises an error if:
+            sample_sheet_file is not an existing file
+            sample_sheet_file contains null byte(s)
+
+    arguments:
+            data_dir -- the directory that has SampleSheet.csv in it
+
+    returns a csv.reader object
+    """
+
+    if path.isfile(sample_sheet_file):
+        csv_file = open(sample_sheet_file, "r")
+        # strip any trailing newline characters from the end of the line
+        # including Windows newline characters (\r\n)
+        csv_lines = [x.rstrip('\n') for x in csv_file]
+        csv_lines = [x.rstrip('\r') for x in csv_lines]
+
+        # open and read file in binary then send it to be parsed by csv's reader
+        csv_reader = reader(csv_lines)
+    else:
+        raise exceptions.SampleSheetError("Sample sheet cannot be parsed as a CSV file because it's not a regular file.",
+                                          sample_sheet_file)
+
+    return csv_reader
