+++ iridauploader/tests/parsers/nextseq/test_validation.py
@@ -15,7 +15,7 @@ class TestValidation(unittest.TestCase):
     def setUp(self):
         print("\nStarting " + self.__module__ + ": " + self._testMethodName)
 
-    @patch("iridauploader.parsers.nextseq.validation.get_csv_reader")
+    @patch("iridauploader.parsers.common.get_csv_reader")
     def test_validate_sample_sheet_no_header(self, mock_csv_reader):
         """
         Given a sample sheet with no header, make sure the correct errors are included in the response
@@ -61,7 +61,7 @@ class TestValidation(unittest.TestCase):
         # Error type should be SampleSheetError
         self.assertEqual(type(res.error_list[0]), SampleSheetError)
 
-    @patch("iridauploader.parsers.nextseq.validation.get_csv_reader")
+    @patch("iridauploader.parsers.common.get_csv_reader")
     def test_validate_sample_sheet_no_data(self, mock_csv_reader):
         """
         Given a sample sheet with no data, make sure the correct errors are included in the response
@@ -112,7 +112,7 @@ class TestValidation(unittest.TestCase):
         self.assertEqual(type(res.error_list[0]), SampleSheetError)
         self.assertEqual(type(res.error_list[1]), SampleSheetError)
 
-    @patch("iridauploader.parsers.nextseq.validation.get_csv_reader")
+    @patch("iridauploader.parsers.common.get_csv_reader")
     def test_validate_sample_sheet_missing_data_header(self, mock_csv_reader):
         """
         Given a sample sheet with no data header, make sure the correct errors are included in the response
@@ -172,7 +172,7 @@ class TestValidation(unittest.TestCase):
         # Error type should be SampleSheetError
         self.assertEqual(type(res.error_list[0]), SampleSheetError)
 
-    @patch("iridauploader.parsers.nextseq.validation.get_csv_reader")
+    @patch("iridauploader.parsers.common.get_csv_reader")
     def test_validate_sample_sheet_valid(self, mock_csv_reader):
         """
         Given a valid sample sheet, test that everything shows as valid
